==> 3._last_fm.ipynb <==
import requests
import pandas as pd
# Configuración para poder visualizar todas las columnas de los DataFrames
pd.set_option('display.max_columns', None) 
import json

1. Para poder conseguir la información requerida en esta fase, usamos la función pd.read_csv, de la biblioteca Pandas: traemos el archivo CSV que obtuvimos de la fase anterior y lo convertirlos en un DataFrame que es almacenado en la variable df_lista_artistas.


df_lista_artistas = pd.read_csv("lista_artistas_musicbrainz.csv")
df_lista_artistas

2. Covertimos el DataFrame en una lista, para poder acceder a la columna que tiene el nombre '0' y así obtener todos los valores de esa columna, o sea, el nombre de los artistas.


artistas = list(df_lista_artistas['0'])
artistas

3. Nos registramos en la página de API Last.fm, y así obtuvimos la ClaveApi (Key), la cual es requerida para hacer la llamada. 

API_KEY = '08f47a34ee1110725e15b70e6e93b478'

- Revisamos y analizamos la documentación proporcionada por la Api, y allí obtuvimos la URL para hacer la llamada. Encontramos que La URL raíz de la API se encuentra en "http://ws.audioscrobbler.com/2.0/" y para realizar la busqueda de la información de los artistas por nombre, la completamos con la URL JSON: "/2.0/?method=artist.search&artist=cher&api_key=SU_CLAVE_API&format=json".

- La documentación también nos indicaba los parámetros: limit, page, artist y Api_key, sin embargo, solo artist y Api_key son los obligatorios y fue los que usamos en nuestra URL.

- Creamos un diccionario con listas vacías, para almacenar la información que obtuvimos del Bucle For
- ...

info_artistas = {'Nombre':[],'Biografia':[], 'Listeners':[], 'Playcount':[], 'Artistas similares':[]}

for artista in artistas:
    url = f'http://ws.audioscrobbler.com/2.0/?method=artist.getinfo&artist={artista}&api_key={API_KEY}&format=json' 

    datos_artista = requests.get(url)
    data = datos_artista.json()
    data

    info_artistas['Nombre'].append(data['artist']['name'])
    info_artistas['Biografia'].append(data['artist']['bio']['content'])
    info_artistas['Listeners'].append(data['artist']['stats']['listeners'])
    info_artistas['Playcount'].append(data['artist']['stats']['playcount'])
    art_similares = []
    for name in data['artist']['similar']['artist']:
        art_similares.append(name['name'])
    info_artistas['Artistas similares'].append(art_similares)

info_artistas


df = pd.DataFrame(info_artistas)

df
