==> 4_update_SQL.ipynb <==
import pandas as pd

**Creación de lista_artistas_id.csv**

A partir del csv lista_artistas_spotify, que contiene una lista de solo nombres de artistas (únicos), vamos a crear un csv que nos facilite la relaciones y consultas en SQL, ya que contendrá un id que será el punto de unión con las tablas que crearemos con los datos de extraídos de Spotify, MusicBraniz y LastFm 

Estos datos serán los que contendrá la tabla **artistas_spotify** en SQL

# Leemos el csv con pandas
lista_artistas_id = pd.read_csv('lista_artistas_spotify.csv')

# #### Esto no hace falta###### la lista ya está con artistas únicos
lista_artistas_id = lista_artistas_id.drop_duplicates()

# Creamos un id_artista, que será único para cada nombre de artista.
# Para ello samos un range 
lista_artistas_id['id_artista'] = range(1, len(lista_artistas_id) + 1)

# Renombrarmos para que la columna que contiene el nombre de artistas, tenga cabecera
lista_artistas_id = lista_artistas_id.rename(columns={'0': 'artista'})

lista_artistas_id

# Y la convertimos en csv para poder insertar los datos en SQL
# Con esto ya tenemos una lista de nombres de artistas y cada uno tiene un id único
lista_artistas_id.to_csv('lista_artistas_id.csv', index=False)

**Creación de spotify_id.csv**

A partir del csv spotify_df_concatenado,vamos a crear otro, que, a parte de los datos resultantes de la extracción y limpieza de los extraídos en la API de Spotify, contendrá un id que identificará al artista y que se corresponderá con el que ya hemos creado en lista_artistas_id.csv. 
Esto nos permitirá relacionar las tablas que crearemos para ambos en SQL 

Estos datos serán los que contendrá la tabla **canciones_spotify** en SQL

# Leemos el archivo csv
spotify = pd.read_csv('spotify_df_concatenado.csv')

spotify

# Vamos a hacer que spotify herede el id de artista lista_artistas_id y, para ello,
# unimos los dos DataFrames en base a la columna 'artista', manteniendo el 
# mismo número de registros de spotify
spotify_id = pd.merge(lista_artistas_id, spotify, on='artista', how='left')

spotify_id


# la columna de artista ya no nos sirve, porque ese dato lo va a suministrar la tabla
# artistas_spotify, así que lo borramos
spotify_id = spotify_id.drop(columns=['artista'])

spotify_id

# Convertimos el archivo en el csv que usaremos para la insercción de datos en SQL
spotify_id.to_csv('spotify_id.csv', index=False)

**Creación de musicbrainz_id.csv**

A partir del csv musicbrainz_df_concatenado, hacemos el mismo procedimiento que en el anterior, para obtener los datos extraídos en la API Music Brainz, que van a incluir ese id que identificará al artista y que se corresponderá con el que ya hemos creado en lista_artistas_id.csv, y nos permita la relación entre las tablas en SQL.

Estos datos serán los que contendrá la tabla **info_artistas_musicbrainz** en SQL

# Leemos el csv que tiene los datos extraidos
musicbrainz_id = pd.read_csv('musicbrainz_df_concatenado.csv')

# Renombramos las cabeceres del Dataframe
musicbrainz_id = musicbrainz_id.rename(columns={'Nombre': 'artista'})
# Y eliminamos la primera columna, que contiene un índice que crea el DataFrame
musicbrainz_id = musicbrainz_id.drop(columns=['Unnamed: 0'])

# Unimos la lista lista_artistas_id y la de musicbrainz_id para conseguir el id de artista
musicbrainz_id = pd.merge(lista_artistas_id, musicbrainz_id, on='artista', how='left')

musicbrainz_id = musicbrainz_id.drop_duplicates()

# la columna de artista ya no nos sirve, porque ese dato lo va a suministrar la tabla
# artistas_spotify, así que lo borramos
musicbrainz_id = musicbrainz_id.drop(columns=['artista'])

# Convertimos el archivo en el csv que usaremos para la insercción de datos en SQL
musicbrainz_id.to_csv('musicbrainz_id.csv', index=False)

musicbrainz_id

**Creación de lastfm_id.csv**

A partir del csv lastfm_df_concatenado, hacemos el mismo procedimiento que en el anterior, para obtener los datos extraídos en la API Last FM, que van a incluir ese id que identificará al artista y que se corresponderá con el que ya hemos creado en lista_artistas_id.csv, y nos permita la relación entre las tablas en SQL.

Estos datos serán los que contendrá la tabla **info_artistas_lastfm** en SQL

# Leemos el csv que contiene los datos
lastfm_id = pd.read_csv('lastfm_df_concatenado.csv')

# Renombramos la columna de artista del DataFrame
lastfm_id = lastfm_id.rename(columns={'Nombre': 'artista'})

# Unimos lista_artistas_id con lastfm_id para añadir el id_artista correspondiente
lastfm_id = pd.merge(lista_artistas_id, lastfm_id, on='artista', how='left')

lastfm_id = lastfm_id.drop_duplicates()

# Eliminamos la columna artista, por no ser necesaria en la futura tabla
lastfm_id = lastfm_id.drop(columns=['artista'])

# Creamos el csv que nos servirá para la insercción de datos en SQL
lastfm_id.to_csv('lastfm_id.csv', index=False)

lastfm_id
