==> 5_Insercion_datos.ipynb <==
!pip install mysql-connector

!pip install mysql-connector-python

!pip install pandas mysql-connector-python

!pip install sqlalchemy 

!pip install pymysql

# Importar librería para la conexión con MySQL
import pandas as pd
import mysql.connector 
from mysql.connector import errorcode

# Importar librerías para manipulación y análisis de datos
import pandas as pd
import numpy as np

# INSERCIÓN DE DATOS DE LA LISTA DE ARTISTAS

# El csv se convierte en un df
df_lista_artistas_id = pd.read_csv('SQL_lista_artistas_id.csv')

# Se convierte en una lista de tuplas para poder insertar los datos a través de python + SQL
tuplas_lista_artistas_id = list(df_lista_artistas_id .itertuples(name=None, index=False))

from sqlalchemy import create_engine

try:
    engine = create_engine('mysql+pymysql://root:AlumnaAdalab@127.0.0.1/music_sp_mb_lfm')

    # Guardar en SQL
    df_lista_artistas_id.to_sql('artistas_spotify', con=engine, if_exists='append', index=False)
    print("Datos insertados exitosamente en la base de datos.")

except Exception as e:
    print("Ocurrió un error al insertar los datos:", e)

# INSERCIÓN DE DATOS DE SPOTIFY

# El csv se convierte en un df
df_spotify_id =  pd.read_csv('SQL_spotify_id.csv')

df_spotify_id = df_spotify_id.drop(df_spotify_id.index[-1])

df_spotify_id

from sqlalchemy import create_engine

try:
    engine = create_engine('mysql+pymysql://root:AlumnaAdalab@127.0.0.1/music_sp_mb_lfm')

    # Guardar en SQL
    df_spotify_id.to_sql('canciones_spotify', con=engine, if_exists='append', index=False)
    print("Datos insertados exitosamente en la base de datos.")

except Exception as e:
    print("Ocurrió un error al insertar los datos:", e)


# INSERCIÓN DE LOS DATOS DE MUSICBRAINZ

# Convertir el csv a dataframe
df_musicbrainz_id = pd.read_csv('SQL_musicbrainz_id.csv')

#Cambio de nombre a las columnas para que no haya error
df_musicbrainz_id.rename(columns={'Nombre': 'id_artista', 'Tipo': 'tipo', 'País de Origen': 'pais_origen', 
                                               'Área de Origen': 'area_origen', 'Inicio de actividad': 'inicio_actividad',
                                                'Fin de actividad': 'fin_actividad', 'Fecha nacimiento': 'fecha_nacimiento',
                                                'Fecha fallecimiento': 'fecha_fallecimiento'}, inplace=True)

print(df_musicbrainz_id)

# Reecuento de los 'NaN' en todas las columnas
nan_counts = df_musicbrainz_id.isna().sum()
print(nan_counts)

# Reemplazar NaN por None en todo el DataFrame
df_musicbrainz_id=df_musicbrainz_id.where(pd.notna(df_musicbrainz_id), None)


# Comprobación 
df_musicbrainz_id

try:
    engine = create_engine('mysql+pymysql://root:AlumnaAdalab@127.0.0.1/music_sp_mb_lfm')

    # Guardar en SQL
    df_musicbrainz_id.to_sql('info_artistas_musicbrainz', con=engine, if_exists='append', index=False)
    print("Datos insertados exitosamente en la base de datos.")

except Exception as e:
    print("Ocurrió un error al insertar los datos:", e)

df_musicbrainz_id = df_musicbrainz_id.drop(df_musicbrainz_id.index[-1])

# INSERCIÓN DE DATOS DE LAST_FM

# Convertir el csv a dataframe
df_lastfm_id = pd.read_csv('SQL_lastfm_id.csv')

 #Cambiar nombres de columnas específicas
df_lastfm_id.rename(columns={'Nombre': 'id_artista_lfm', 'Biografia': 'biografia', 'Listeners': 'listeners', 
                                               'Playcount': 'playcount', 'Artistas similares': 'artistas_similares'
                                                }, inplace=True)

print(df_lastfm_id)

try:
    engine = create_engine('mysql+pymysql://root:AlumnaAdalab@127.0.0.1/music_sp_mb_lfm')

    # Guardar en SQL
    df_lastfm_id.to_sql('info_artistas_lastfm', con=engine, if_exists='append', index=False)
    print("Datos insertados exitosamente en la base de datos.")

except Exception as e:
    print("Ocurrió un error al insertar los datos:", e)


#Cerrar conexión
engine.dispose()  # Esto libera todas las conexiones
print("Conexión cerrada.")
